feat: add routes and folder organization
